Move runtime interface
